Better handling of sign-out.
Now clear some cached values so that they'll be refreshed with the new
user's values. Affects the sign-in email, TB-Loader ID, and custom
greeting.

(Previously, if the user signed out and back in with a different ID, any
new SRNs would still be assigned with the old ID. They would have still
been unique, because the blocks are allocated per-user, per-device.)